pyproject.toml
[project]
name = "alpha-vantage-mcp"
version = "0.1.0"
description = "AWS Lambda function with MCP CLI support"
requires-python = ">=3.13"
dependencies = [
    "awslabs-mcp-lambda-handler>=0.1.8",
    "loguru>=0.7.3",
    "mcp>=1.12.3",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
include = [
    "lambda_function.py",
    "requirements.txt"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]
